Change Syringe Runtime to avoid using a class
The Runtime really only neeeds some very basic data, primarily a list of target
function pointers, and the addresses of their alternate implementations.
Additionally, it will be necessary to provide some locking/synchronization
mechanisms for each implementation. A class interface is not necesary, and
complicates the architecture of the runtime.