Change Syringe Payload Attribute to be only in clang
Origianlly I had thought of these attributes as more or less booleans,
and implemented them inside of LLVM, since they did not need any parameters.

To facilitate configuration files, it will be desireable to pass parametes to
the SyringePayload attribute, and also allow the config to be done 100% in
source code.