Update Ctor generation to support direct calls
We now use a smaller data structure for metadata, and also
need to bitcast some parameters for method calls.

In particular, the this parameter for C++ methods will be the
wrong type, and thus trigger an assertion error. We can either
add checks for Syringe attributes to the assertions, or bitcast
in these cases.